pyproject.toml
[project]
name = "content-system"
version = "0.1.0"
description = "API for movie information upload and retrieval"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.4",
    "djangorestframework>=3.15.2",
    "pandas>=2.2.3",
    "pytest>=8.3.4",
    "ruff>=0.8.3",
]
